requirements.txt
Package         Version   
--------------- ----------
beautifulsoup4  4.8.2     
certifi         2019.11.28
chardet         3.0.4     
idna            2.8       
numpy           1.18.1    
pandas          1.0.1     
pip             19.0.3    
python-dateutil 2.8.1     
pytz            2019.3    
requests        2.22.0    
setuptools      40.8.0    
six             1.14.0    
soupsieve       1.9.5     
urllib3         1.25.8    
